jingdong_captcha_break/requirements.txt
selenium==3.141.0
numpy==1.15.4
opencv_python==3.4.3.18
Pillow==5.3.0
